# docs/storoom.api-docs.yaml
openapi: 3.0.1
info:
  title: storoom
  version: v1
servers:
- url: https://localhost:8080/
security:
- Bearer Authentication: []
paths:
...
  /owners/{id}:
    put:
      tags:
      - Owners
      description: Update owner.
      operationId: updateOwner
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OwnerInfoDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OwnerReadDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "404":
          description: Resource not found.
        "409":
          description: Request action conflicts with the current state of things.
    delete:
      tags:
      - Owners
      description: Delete owner.
      operationId: deleteOwner
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OwnerReadDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "404":
          description: Resource not found.
        "409":
          description: Request action conflicts with the current state of things.
  /orders/{id}:
    put:
      tags:
      - Orders
      description: Update order
      operationId: updateOrder
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderInfoDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OrderFullDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "404":
          description: Resource not found.
        "409":
          description: Request action conflicts with the current state of things.
  /manufacturers/{id}:
    get:
      tags:
      - Manufacturers
      description: Get manufacturer by id.
      operationId: getManufacturer
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ManufacturerDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "404":
          description: Resource not found.
    put:
      tags:
      - Manufacturers
      description: Update manufacturer's name.
      operationId: updateManufacturerName
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManufacturerNameDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ManufacturerDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "404":
          description: Resource not found.
        "409":
          description: Request action conflicts with the current state of things.
    delete:
      tags:
      - Manufacturers
      description: Delete manufacturer.
      operationId: deleteManufacturer
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ManufacturerDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "404":
          description: Resource not found.
        "409":
          description: Request action conflicts with the current state of things.
...
  /locations/{locationId}/owners/{ownerId}:
    put:
      tags:
      - Locations
      description: Add owner to location.
      operationId: addOwnerToLocation
      parameters:
      - name: locationId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: ownerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LocationReadDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "404":
          description: Resource not found.
        "409":
          description: Request action conflicts with the current state of things.
    delete:
      tags:
      - Locations
      description: Delete owner from location.
      operationId: deleteOwnerFromLocation
      parameters:
      - name: locationId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: ownerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/LocationReadDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "404":
          description: Resource not found.
        "409":
          description: Request action conflicts with the current state of things.
...
  /owners:
    get:
      tags:
      - Owners
      description: Get all owners.
      operationId: getOwners
      parameters:
      - name: page
        description: number of page
      - name: size
        description: size of page
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OwnerCompactDTO'
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
    post:
      tags:
      - Owners
      description: Add owner.
      operationId: addOwner
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OwnerDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OwnerReadDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "409":
          description: Request action conflicts with the current state of things.
  /owners/{ownerId}/locations/{locationId}:
    post:
      tags:
      - Owners
      description: Add location to owner.
      operationId: addLocationToOwner
      parameters:
      - name: ownerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: locationId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OwnerReadDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "404":
          description: Resource not found.
        "409":
          description: Request action conflicts with the current state of things.
    delete:
      tags:
      - Owners
      description: Delete location from owner.
      operationId: deleteLocationFromOwner
      parameters:
      - name: ownerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: locationId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OwnerReadDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "404":
          description: Resource not found.
        "409":
          description: Request action conflicts with the current state of things.
  /orders:
    get:
      tags:
      - Orders
      description: Get orders
      operationId: getOrders
      parameters:
      - name: userId
        in: query
        description: (optional) userId to search for
        required: false
        schema:
          type: string
          format: uuid
      - name: page
        description: number of page
      - name: size
        description: size of page
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageOrderDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "404":
          description: Resource not found.
    post:
      tags:
      - Orders
      description: Create order
      operationId: createOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OrderFullDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "404":
          description: Resource not found.
        "409":
          description: Request action conflicts with the current state of things.
  /orders/{id}/finish:
    post:
      tags:
      - Orders
      description: Finish order
      operationId: finishOrder
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/OrderFullDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
        "404":
          description: Resource not found.
        "409":
          description: Request action conflicts with the current state of things.
  /manufacturers:
    get:
      tags:
      - Manufacturers
      description: Get all manufacturers.
      operationId: getManufacturers
      parameters:
      - name: page
        description: number of page
      - name: size
        description: size of page
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ManufacturerDTO'
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
    post:
      tags:
      - Manufacturers
      description: Add manufacturer.
      operationId: createManufacturer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManufacturerDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ManufacturerDTO'
        "400":
          description: Request is incorrect.
        "401":
          description: You are not authorized to access the resource.
        "403":
          description: You are forbidden to access the resource.
...
components:
  schemas:
...
    LocationDTO:
      type: object
      properties:
        address:
          type: string
        locationType:
          type: string
          enum:
          - stand
          - warehouse
          - main_office
        ownerIds:
          type: array
          items:
            type: string
            format: uuid
    LockDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        manufacturer:
          type: string
          format: uuid
    UnitFullDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sizeX:
          type: integer
          format: int32
        sizeY:
          type: integer
          format: int32
        sizeZ:
          type: integer
          format: int32
        unitType:
          type: string
          enum:
          - S
          - M
          - L
          - XL
        status:
          type: string
          enum:
          - dead
          - available
          - occupied
          - pending
        location:
          $ref: '#/components/schemas/LocationDTO'
        lock:
          $ref: '#/components/schemas/LockDTO'
...
    OwnerInfoDTO:
      type: object
      properties:
        name:
          type: string
        companyId:
          type: string
          format: uuid
    LocationCompactDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        address:
          type: string
        locationType:
          type: string
          enum:
          - stand
          - warehouse
          - main_office
    OwnerReadDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        companyId:
          type: string
          format: uuid
        locations:
          type: array
          items:
            $ref: '#/components/schemas/LocationCompactDTO'
    OrderInfoDTO:
      type: object
      properties:
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        finishedTime:
          type: string
          format: date-time
        unitId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
    OrderFullDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        finishedTime:
          type: string
          format: date-time
        status:
          type: string
          enum:
          - active
          - finished
        unit:
          $ref: '#/components/schemas/UnitFullDTO'
        user:
          $ref: '#/components/schemas/UserDTO'
    UserDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
    ManufacturerNameDTO:
      type: object
      properties:
        name:
          type: string
    ManufacturerDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
...
    LocationReadDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        address:
          type: string
        locationType:
          type: string
          enum:
          - stand
          - warehouse
          - main_office
        owners:
          type: array
          items:
            $ref: '#/components/schemas/OwnerCompactDTO'
    OwnerCompactDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        companyId:
          type: string
          format: uuid
...
    OwnerDTO:
      type: object
      properties:
        name:
          type: string
        companyId:
          type: string
          format: uuid
        locationIds:
          type: array
          items:
            type: string
            format: uuid
    OrderDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        finishedTime:
          type: string
          format: date-time
        status:
          type: string
          enum:
          - active
          - finished
        unitId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
...
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/SortObject'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...
    PageOrderDTO:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/OrderDTO'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        pageable:
          $ref: '#/components/schemas/PageableObject'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean
...
  securitySchemes:
    Bearer Authentication:
      type: http
      scheme: bearer
      bearerFormat: JWT
    Basic Authentication:
      type: http
      scheme: basic